Refacto Quickalgo executor and add check end condition at every turn
