+++ frontend/stepper/api.ts
@@ -27,6 +27,7 @@ import {TaskSubmissionResultPayload} from "../submission/submission";
 import {QuickAlgoLibrary} from '../task/libs/quickalgo_library';
 import {CodecastPlatform} from './platforms';
 import {LibraryTestResult} from '../task/libs/library_test_result';
+import {createQuickAlgoLibraryExecutor} from './quickalgo_executor';
 
 export interface QuickalgoLibraryCall {
     module: string,
@@ -478,150 +479,6 @@ function inUserCode(stepperState: StepperState) {
         return true;
     } else {
         return !!stepperState.programState.control.node[1].begin;
-    }
-}
-
-export function createQuickAlgoLibraryExecutor(stepperContext: StepperContext) {
-    return async (module: string, action: string, args: any[], callback?: Function) => {
-        log.getLogger('quickalgo_executor').debug('[quickalgo_executor] call quickalgo', module, action, args, callback);
-        let libraryCallResult;
-        const context = stepperContext.quickAlgoContext;
-
-        // Wait that the context has finished all previous animations
-        if (stepperContext.waitPreviousAnimations) {
-            log.getLogger('quickalgo_executor').debug('[quickalgo_executor] before ready check');
-            await new Promise((resolve) => {
-                context.executeWhenReady(resolve);
-            });
-        }
-
-        log.getLogger('quickalgo_executor').debug('[quickalgo_executor] ready for call');
-
-        if (stepperContext.state) {
-            log.getLogger('quickalgo_executor').debug('[quickalgo_executor] stepper context before', stepperContext.state.contextState);
-        }
-
-        if (stepperContext.quickAlgoCallsLogger) {
-            const quickAlgoLibraryCall: QuickalgoLibraryCall = {module, action, args};
-            stepperContext.quickAlgoCallsLogger(quickAlgoLibraryCall);
-            log.getLogger('quickalgo_executor').debug('[quickalgo_executor] call quickalgo calls logger', module, action, args);
-        }
-
-        const makeLibraryCall = () => {
-            return new Promise((resolve, reject) => {
-                let callbackArguments = [];
-                let result = context[module][action].apply(context, [...args, function (a) {
-                    log.getLogger('quickalgo_executor').debug('[quickalgo_executor] receive callback', arguments);
-                    callbackArguments = [...arguments];
-                    let argumentResult = callbackArguments.length ? callbackArguments[0] : undefined;
-                    log.getLogger('quickalgo_executor').debug('[quickalgo_executor] set result', argumentResult);
-                    resolve(argumentResult);
-                }]);
-
-                log.getLogger('quickalgo_executor').debug('[quickalgo_executor] MODULE RESULT', result);
-                if (Symbol.iterator in Object(result)) {
-                    iterateResult(result)
-                        .then((argumentResult) => {
-                            log.getLogger('quickalgo_executor').debug('[quickalgo_executor] returned element', module, action, argumentResult);
-                            // Use context.waitDelay to transform result to primitive when the library uses generators
-                            context.waitDelay(resolve, argumentResult);
-                        })
-                        .catch(reject)
-                }
-            });
-        }
-
-        const iterateResult = async (result) => {
-            let lastResult;
-            while (true) {
-                /* Pull the next effect from the builtin's iterator. */
-                const {done, value} = result.next();
-                log.getLogger('quickalgo_executor').debug('[quickalgo_executor] ITERATOR RESULT', module, action, done, value);
-                if (done) {
-                    return value;
-                }
-
-                const name = value[0];
-                if (name === 'interact') {
-                    log.getLogger('quickalgo_executor').debug('[quickalgo_executor] ASK FOR INTERACT', stepperContext.interactAfter);
-                    lastResult = await stepperContext.interactAfter({...(value[1] || {}), progress: false});
-                    log.getLogger('quickalgo_executor').debug('[quickalgo_executor] last result', lastResult);
-                } else if (name == 'put') {
-                    log.getLogger('quickalgo_executor').debug('[quickalgo_executor] ask put dispatch', value[1]);
-                    await stepperContext.dispatch(value[1]);
-                }
-            }
-        }
-
-        log.getLogger('quickalgo_executor').debug('[quickalgo_executor] before make async library call', {module, action});
-        let hideDisplay = false;
-        let previousDelay = context.getDelay();
-        try {
-            if ('main' === stepperContext.environment) {
-                if ('running' === stepperContext.taskDisplayNoneStatus) {
-                    // context.display = false;
-                    hideDisplay = true;
-                    // context.needsRedrawDisplay = true;
-                    // } else if ('end' === stepperContext.taskDisplayNoneStatus) {
-                    context.changeDelay(0);
-                    stepperContext.taskDisplayNoneStatus = null;
-                }
-            }
-
-            libraryCallResult = await makeLibraryCall();
-            log.getLogger('quickalgo_executor').debug('[quickalgo_executor] after make async lib call', libraryCallResult);
-
-            if (hideDisplay) {
-                // context.display = true;
-            // } else {
-                context.changeDelay(previousDelay);
-            }
-
-            // Leave stepperThrottleDisplayDelay ms before displaying again the context
-            if (!stepperContext.taskDisplayNoneStatus) {
-                stepperContext.taskDisplayNoneStatus = 'running';
-                setTimeout(() => {
-                    stepperContext.taskDisplayNoneStatus = 'end';
-                }, stepperThrottleDisplayDelay);
-            }
-        } catch (e: unknown) {
-            log.getLogger('quickalgo_executor').debug('[quickalgo_executor] context error 2', e);
-            if (hideDisplay) {
-                // context.display = true;
-            // } else {
-                context.changeDelay(previousDelay);
-            }
-
-            await stepperContext.dispatch({
-                type: CompileActionTypes.StepperInterrupting,
-            });
-
-            Codecast.runner.stop();
-
-            await stepperContext.dispatch(stepperExecutionError(LibraryTestResult.fromString(String(e)), false));
-        }
-
-        log.getLogger('quickalgo_executor').debug('[quickalgo_executor] after make async library call', libraryCallResult);
-
-        const newState = getCurrentImmerState(context.getInnerState());
-        log.getLogger('quickalgo_executor').debug('[quickalgo_executor] NEW LIBRARY STATE', newState);
-
-        if (stepperContext.state) {
-            stepperContext.state = {
-                ...stepperContext.state,
-                contextState: newState,
-            };
-            log.getLogger('quickalgo_executor').debug('[quickalgo_executor] stepper context after', stepperContext.state.contextState);
-        }
-
-        if (callback) {
-            log.getLogger('quickalgo_executor').debug('[quickalgo_executor] call callback arguments', libraryCallResult);
-            callback(libraryCallResult);
-        }
-
-        log.getLogger('quickalgo_executor').debug('[quickalgo_executor] return library call result', libraryCallResult);
-
-        return libraryCallResult;
     }
 }
