+++ frontend/task/task_slice.ts
@@ -5,6 +5,7 @@ import ProcessingFixture from './fixtures/test_processing';
 import BarcodeFixture from './fixtures/test_barcode';
 import QuickPiFixture from './fixtures/quickpi_testbed';
 import QuickPiBuzzerFixture from './fixtures/quickpi_buzzer';
+import QuickPiDirectionFixture from './fixtures/quickpi_direction';
 import SokobanFixture from './fixtures/11_variable_08_sokoban';
 import StringsLoginFixture from './fixtures/14_strings_01_login';
 import DatabaseFixture from './fixtures/test_database';
@@ -21,6 +22,7 @@ const availableTasks = {
     turtle: TurtleFixture,
     quickpi: QuickPiFixture,
     buzzer: QuickPiBuzzerFixture,
+    direction: QuickPiDirectionFixture,
     processing: ProcessingFixture,
     printer: StringRotationFixture,
     barcode: BarcodeFixture,
